requirements.txt
streamlit>=1.24.0
pandas>=1.5.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
openai>=1.0.0 